# valo-mapper-api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
  handlers.RSOCallbackResponse:
    properties:
      customToken:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.RSORequest:
    properties:
      code:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      firebaseUid:
        type: string
      id:
        type: integer
      isSubscribed:
        type: boolean
      name:
        type: string
      rsoLinkedAt:
        type: string
      rsoSubjectId:
        type: string
      subscriptionEndedAt:
        type: string
      tourCompleted:
        type: boolean
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: API for ValoMapper backend services.
  title: ValoMapper API
  version: "1.0"
paths:
  /api/auth/rso/callback:
    post:
      consumes:
      - application/json
      description: Exchanges a Riot authorization code and returns a Firebase custom
        token plus user data.
      parameters:
      - description: RSO authorization code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RSORequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RSOCallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Exchange RSO code for Firebase custom token
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Firebase bearer token in the form "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
  InternalAPIKey:
    description: Internal service API key.
    in: header
    name: X-Internal-API-Key
    type: apiKey
swagger: "2.0"